change int to float64 in Account struct and according methods

+++ internal/domain/entities/account.go
@@ -1,6 +1,6 @@
 package entities
 
 type Account struct {
-	Id      string `json:"id"`
-	Balance int    `json:"balance"`
+	Id      string  `json:"id"`
+	Balance float64 `json:"balance"`
 }
