+++ internal/ports/input/account.go
@@ -7,6 +7,6 @@ import (
 
 type AccountService interface {
 	Create(ctx context.Context, account entities.Account) error
-	GetBalance(ctx context.Context, account entities.Account) (balance int, err error)
-	Transfer(ctx context.Context, from entities.Account, to entities.Account, amount int) error
+	GetBalance(ctx context.Context, account entities.Account) (balance float64, err error)
+	Transfer(ctx context.Context, from entities.Account, to entities.Account, amount float64) error
 }
